(chore) refactor Login and Register folder structure
